+++ ecommerce/frontend/src/components/layout/Content.js
@@ -1,8 +1,8 @@
 import { Route, Switch, Link } from 'react-router-dom'
 import { Layout } from 'antd'
-import { LoginPage } from '../LoginPage'
-import { SignupPage } from '../SignupPage'
-import { HomePage } from '../HomePage'
+import { LoginPage } from '../pages/LoginPage'
+import { SignupPage } from '../pages/SignupPage'
+import { HomePage } from '../pages/HomePage'
 
 export const Content = () => {
     return (
